Working test against FS
